refactor: better architecture
