+++ bot.py
@@ -15,9 +15,9 @@ from telegram.ext import (
     ContextTypes
 )
 
-from data.flight_info import FlightInfo
-from data.flight_status import FlightStatus
-from data.ticket_booking import TicketBooking
+from services.flight_info import FlightInfo
+from services.flight_status import FlightStatus
+from services.ticket_booking import TicketBooking
 from nlp.casual_dialog import CasualDialogHandler
 from nlp.intent_model import IntentClassifier
 from nlp.ner import extract_city, replace_placeholders
