+++ handlers/commands.py
@@ -1,9 +1,9 @@
 import logging
 from telegram import Update
 from telegram.ext import ContextTypes
-from data.flight_info import FlightInfo
-from data.flight_status import FlightStatus
-from data.ticket_booking import TicketBooking
+from services.flight_info import FlightInfo
+from services.flight_status import FlightStatus
+from services.ticket_booking import TicketBooking
 from handlers.buttons import ButtonHandler
 
 logger = logging.getLogger(__name__)
